+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/index.tsx
@@ -34,7 +34,7 @@ import { AccountSettings } from './views/account_settings';
 import { SourcesRouter } from './views/content_sources';
 import { SourceAdded } from './views/content_sources/components/source_added';
 import { SourceSubNav } from './views/content_sources/components/source_sub_nav';
-import { PersonalDashboardLayout } from './views/content_sources/private_sources_layout';
+import { PersonalDashboardLayout } from './views/content_sources/personal_dashboard_layout';
 import { ErrorState } from './views/error_state';
 import { GroupsRouter } from './views/groups';
 import { GroupSubNav } from './views/groups/components/group_sub_nav';
