+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/views/content_sources/index.ts
@@ -7,4 +7,4 @@
 
 export { Overview } from './components/overview';
 export { SourcesRouter } from './sources_router';
-export { PersonalDashboardLayout } from './private_sources_layout';
+export { PersonalDashboardLayout } from './personal_dashboard_layout';
